refactor: make page responsive
make page responsive to fit different screen sizes

<finishes>
