+++ src/__tests__/component/categories.test.js
@@ -1,7 +1,7 @@
 import { shallow } from 'enzyme';
 import React from 'react';
 import toJson from 'enzyme-to-json';
-import CategoryComponent from '../../components/Category.jsx';
+import CategoryComponent from '../../components/Categories.jsx';
 
 jest.mock('../../api/NewsApi', () => ({
   getData: () => Promise.resolve('getPromise')
