+++ src/style/mainStyle.scss
@@ -15,6 +15,7 @@ $grey: grey;
 $source_border: #dfe3ee;
 $header_color: #333;
 $headerText: #9d9d9d;
+$warningText: indianred;
 
 //fonts
 $font_0: "Roboto";
@@ -51,13 +52,13 @@ a {
   margin-left: 40vw;
   margin-top: 20vw;
 }
+.warningText{
+  color: $warningText;
+}
 .errorMsg{
-  position: fixed;
-  width: 35vw;
-  font-size: 1.4vw;
-  margin-top: -49vw;
-  margin-left: 63vw;
+  font-size: 89%;
   text-align: center;
+  width: 40vw;
 }
 .spin{
   margin-left: 40%;
@@ -90,31 +91,37 @@ a {
 }
 .loginBackgroundImg{
   background-image: $url_0;
-  height: 55vw;
-  width: 60vw;
+  height: 70vh;
+  width: 57vw;
   margin-left: 3vw;
 	background-repeat: no-repeat;
 	background-position: center;
 	background-size: cover;
 }
 .loginHeader{
-  font-family: "Comic Sans MS", "Comic Sans", cursive;
-  padding: 1vw;
   border-bottom: 2px solid;
   box-shadow: 10px 10px 5px $color_tapa_approx;
+  font-family: "Comic Sans MS", "Comic Sans", cursive;
+  font-size: 140%;
+  padding: 1vw;
 }
 .all-available-login-button{
-  position: absolute;
-  margin-top: -40vw;
-  margin-left: 68vw;
-  margin-right: 2vw;
+  margin-top: 20%;
   border: 2px solid;
   padding: 3vw;
-  padding-bottom: 8vw;
   border-radius: 2vw;
   box-shadow: 10px 10px 5px #888888;
 }
-
+@media screen and (max-width: 991px){
+  .loginBackgroundImg{
+    visibility: hidden;
+  }
+  .displayMsg{
+    margin-left: 20vw;
+    margin-top: -70vh;
+    position: absolute
+  }
+}
 .logoutBtn {
   font-size: 150%;
   font-weight: bold;
@@ -172,13 +179,12 @@ p {
 	overflow-y: scroll;
 }
 .outerSourceBorder{
-  width: 21%;
+  width: 21% !important;
   min-height: 51vh;
   max-height: 51vh;
   padding-bottom: 1px;
 }
 .outerArticleCountainer {
-  margin-left: 26%;
   max-height: 71vh;
   min-height: 71vh;
   overflow-y: scroll;
@@ -232,7 +238,6 @@ p {
   margin-left: 3vw;
 }
 .footer{
-  padding-top: 1%;
   position: relative;
   text-align: left;
   margin-left: 3vw;
@@ -301,10 +306,6 @@ p {
 		padding-right: 0;
 		padding-left: 0;
 	}
-	.footer {
-		padding-right: 0;
-		padding-left: 0;
-	}
 }
 .navbar2 {
 	margin-bottom: 1vw;
@@ -320,12 +321,14 @@ p {
 .sortByFeature {
 	font-weight: bold;
   font-size: 100%;
+  margin-top: -2vh;
 }
 .dropdown-toggle{
   color: $color_tapa_approx;
 }
 .login-page{
-  background-color: $mercury;
+  // background-color: $mercury;
+  margin: 0 3% 0 3%;
 }
 .loginBtn--google {
 	font-family: $font_0, $font_1, $font_2, $font_3;
@@ -358,15 +361,14 @@ p {
 .loginBtn {
 	box-sizing: border-box;
 	position: relative;
-	margin-top: 3em;
-	margin-left: 3em;
+	margin-top: 2.6em;
 	padding: 0 15px 0 46px;
 	border: none;
 	text-align: left;
 	line-height: 34px;
 	white-space: nowrap;
 	border-radius: 0.2em;
-	font-size: 16px;
+	font-size: 2.3vh;
 	color: $white;
 	&:before {
 		content: "";
@@ -387,12 +389,24 @@ p {
 .disableClick {
   pointer-events: none;
 }
+.articleSourceContainer{
+  padding-bottom: 0px;
+  margin-bottom: 0px;
+}
 
+.navbar-collapse{
+  margin-top: -2vh;
+  width: 92.6vw;
+  margin-left: 1vw !important;
+}
 /*
  * Off Canvas
  * --------------------------------------------------
  */
 @media screen and (max-width: 767px) {
+  .articleSourceContainer{
+    margin-left: -1%;
+  }
   .row-offcanvas {
     position: relative;
     -webkit-transition: all .25s ease-out;
@@ -453,34 +467,44 @@ p {
   }
 }
 @media (min-width: 767px){
-   .Source-Container {
+  .Source-Container {
     max-height: 52vh;
   }
 }
 @media (max-width: 767px){
   .outerArticleCountainer {
-    margin-left: 30%;
-    width: 67%;
+    width: 74%;
   }
   .show-sources {
     margin-left: 30%;
   }
+  .outerSourceBorder {
+    width: 0%;
+    min-height: 51vh;
+    max-height: 51vh;
+    padding-bottom: 1px;
+  }
+}
 
   @media (min-width: 900px){
     .outerArticleCountainer {
       margin-left: 30%;
       width: 67%;
     }
-  }
   .show-sources {
     margin-left: 30%;
-  }905
+  }
   .outerSourceBorder{
     background-color: #f5f5f5;
     width: 28%;
     min-height: 72vh;
     max-height: 51vh;
     padding-bottom: 1px;
+  }
+}
+@media screen and (max-width: 889px) and (min-width: 768px){
+  .categoryList{
+    font-size: 77% !important;
   }
 }
 @media (min-width: 1366px){
@@ -492,7 +516,25 @@ p {
   .navbar-default {
     display: inline;
   }
+  .text-muted {
+    display: none;
+  }
 }
 li > a{
   font-size: 100%;
+}
+@media (min-width: 1200px){
+  .container {
+    width: 90vw;
+  }
+}
+@media (max-height: 680px){
+  .Source-Container {
+    max-height: 54vh;
+  }
+}
+.text-muted {
+  padding-bottom: -9%;
+  margin-bottom: -4%;
+  margin-left: 71%;
 }
